Sphere object, Noise + Vertical Variation
I made a new sphere object and moved the UVs around a little. The top and bottom is still kinda weird and there is a seam, but it's much better than the default sphere. I also started generating noise for the planet and hooked it up to a gradient to choose the colors based on the "height" of the noise. Lastly I put in an influencer that makes it higher towards the poles to give the imitation of ice.